# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "System manage music",
        "title": "Music API REST",
        "termsOfService": "https://juancx.site/",
        "contact": {
            "name": "API Support",
            "email": "juanm.campos@unas.edu.pe"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:4021",
    "basePath": "/",
    "paths": {
...
        "/api/v1/songsplayed": {
            "put": {
                "description": "Método que permite Actualiza una instancia del objeto SongsPlayed en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsPlayed"
                ],
                "summary": "Actualiza una instancia de SongsPlayed",
                "parameters": [
                    {
                        "description": "Datos para actualizar SongsPlayed",
                        "name": "RequestSongsPlayed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/songs_played.RequestSongsPlayed"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    }
                }
            },
            "post": {
                "description": "Método que permite crear una instancia del objeto SongsPlayed en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsPlayed"
                ],
                "summary": "Crear una instancia de SongsPlayed",
                "parameters": [
                    {
                        "description": "Datos para crear SongsPlayed",
                        "name": "RequestSongsPlayed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/songs_played.RequestSongsPlayed"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    }
                }
            },
            "delete": {
                "description": "Método que permite eliminar una instancia del objeto SongsPlayed en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsPlayed"
                ],
                "summary": "Elimina una instancia de SongsPlayed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SongsPlayed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    }
                }
            }
        },
        "/api/v1/songsplayed/": {
            "get": {
                "description": "Método que permite obtener todas las instancias del objeto SongsPlayed en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsPlayed"
                ],
                "summary": "Obtiene todas las instancias de SongsPlayed",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    }
                }
            }
        },
        "/api/v1/songsplayed/{id}": {
            "get": {
                "description": "Método que permite obtener una instancia del objeto SongsPlayed en la base de datos por su id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SongsPlayed"
                ],
                "summary": "Obtiene una instancia de SongsPlayed por su id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SongsPlayed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/songs_played.ResponseSongsPlayed"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "songs_played.RequestSongsPlayed": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "song": {
                    "type": "integer"
                },
                "user": {
                    "type": "integer"
                }
            }
        },
        "songs_played.ResponseSongsPlayed": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}